refactor!: remove home-manager input
Previously this flake provided home-manager (which could be overridden).
Instead, the flake no longer has a direct dependency on home-manager.
Consuming flakes that wish to use `flakeModule.home` must provide it in
their own inputs. If the name is not `home-manager`, then they must
specify `config-parts.home.home-manager` with the flake input they've
chosen.

I would prefer to keep providing home-manager because I think overriding
the input is more ergonomic and sensical. However, this kind of
architecture is the only way to provide extra modules without inflating
consuming flake lockfiles. This will be more relevant when the
`nix-darwin` module is created as it will allow nixos users to use the
flake without pulling in `nix-darwin` similarly to this change for
`home-manager`.